+++ rus/libdict/capsheme.h
@@ -17,8 +17,8 @@
 
 namespace LIBMORPH_NAMESPACE
 {
-  unsigned  GetCapScheme( unsigned char*  output, unsigned    outlen,
-                          const char*     srctop, unsigned    srclen = (unsigned)-1 );
+  unsigned  GetCapScheme( unsigned char*  output, size_t      outlen,
+                          const char*     srctop, size_t      srclen = (size_t)-1 );
   unsigned  GetMinScheme( unsigned        minCap, const char* lpword = 0, unsigned nparts = 0 );
   char*     SetCapScheme( char*           pszstr, unsigned    scheme );
